Portfolio in Progress 7
Working on grid and trying to get the scroll wheel to work with artwork. May change how the artwork page works to Alison's style of showing work.

BKG needs to either change or be changed. Need to link contact page to email. Sent this version to family members to try